refactor statement execution
