+++ modules/fbs-core/web/src/app/page-components/sql-playground/sql-playground.component.ts
@@ -7,6 +7,8 @@ import { AuthService } from "src/app/service/auth.service";
 import { SqlPlaygroundService } from "src/app/service/sql-playground.service";
 import { Table } from "src/app/model/sql_playground/Table";
 import { Constraint } from "src/app/model/sql_playground/Constraint";
+import { delay, retryWhen } from "rxjs/operators";
+import { MatSnackBar } from "@angular/material/snack-bar";
 
 /**
  * This component is for the sql playground
@@ -20,6 +22,7 @@ export class SqlPlaygroundComponent implements OnInit {
   constructor(
     private titlebar: TitlebarService,
     private authService: AuthService,
+    private snackbar: MatSnackBar,
     private sqlPlaygroundService: SqlPlaygroundService
   ) {}
 
@@ -83,5 +86,48 @@ export class SqlPlaygroundComponent implements OnInit {
       .subscribe((result) => {
         this.triggers = result;
       });
+  }
+
+  submitStatement(statement: string) {
+    this.isQueryPending = true;
+    const token = this.authService.getToken();
+
+    this.sqlPlaygroundService
+      .submitStatement(token.id, this.activeDb, statement)
+      .subscribe(
+        (result) => {
+          this.getResultsbyPolling(result.id);
+        },
+        (error) => {
+          console.error(error);
+          this.snackbar.open(
+            "Beim Versenden ist ein Fehler aufgetreten. Versuche es später erneut.",
+            "OK",
+            { duration: 3000 }
+          );
+          this.isQueryPending = false;
+        }
+      );
+  }
+
+  private getResultsbyPolling(rId: number) {
+    const token = this.authService.getToken();
+
+    this.sqlPlaygroundService
+      .getResults(token.id, this.activeDb, rId)
+      .pipe(
+        retryWhen((err) => {
+          return err.pipe(delay(1000));
+        })
+      )
+      .subscribe(
+        (res) => {
+          // emit if success
+          this.isQueryPending = false;
+          this.resultset = res;
+        },
+        () => {}, //handle error
+        () => console.log("Request Complete")
+      );
   }
 }
